[merge] Move cmap code to merge.cmap
